+++ src/app/dataset-view/additional-components/dataset-settings-component/dataset-settings.component.spec.ts
@@ -175,7 +175,10 @@ describe("DatasetSettingsComponent", () => {
 
         tick();
 
-        expect(deleteDatasetSpy).toHaveBeenCalledOnceWith(component.datasetBasics.id);
+        expect(deleteDatasetSpy).toHaveBeenCalledOnceWith(
+            component.datasetBasics.id,
+            component.datasetBasics.owner.accountName,
+        );
 
         flush();
     }));